requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==2.5.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.0.12
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
flake8==3.9.2
flake8-docstrings==1.6.0
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==4.6.0
lru-dict==1.1.7
mccabe==0.6.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pycodestyle==2.7.0
pycryptodome==3.14.1
pydocstyle==6.1.1
pyflakes==2.3.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==6.2.5
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
rlp==2.0.1
six==1.16.0
snowballstemmer==2.2.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
varint==1.0.2
web3==5.29.2
websockets==9.1
yarl==1.7.2
